pyproject.toml
[project]
name = "airflow-pipelines"
version = "0.1.0"
description = ""
authors = [
    { name = "Almaz Khayfullin", email = "ah@carbox.tech" }
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "apache-airflow==2.9.3",
    "airflow-clickhouse-plugin==1.4.0",
    "clickhouse-driver==0.2.6",
    "clickhouse-connect (>=0.8.17,<0.9.0)",
    "ijson (>=3.4.0,<4.0.0)",
    "xmltodict (>=1.0.2,<2.0.0)",
    "boto3 (==1.26.165)",
    "soda-core-mysql-utf8-hotfix (==3.5.5.post1)",
    "soda-core (==3.5.5)",
    "mysql-connector-python (==8.0.33)",
    "gspread (>=6.2.1,<7.0.0)",
    "google-auth (>=2.43.0,<3.0.0)",
    "botocore (==1.29.165)",
    "pyarrow (>=22.0.0,<23.0.0)",
    "yfinance (>=0.2.43,<0.3.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
black = ">=24.0.0"
isort = ">=5.12.0"
flake8 = ">=6.1.0"
ipykernel = "^6.29.5"
notebook = "^7.4.3"
jupyter = "^1.1.1"

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.isort]
profile = "black"

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
